+++ packages/query-module-pagination/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rym-lib/query-module-pagination
 
+## 1.2.2
+
+### Patch Changes
+
+- update nakadachi, query-module
+
 ## 1.2.1
 
 ### Patch Changes