+++ packages/rdb-command/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rym-lib/rdb-command
 
+## 1.2.2
+
+### Patch Changes
+
+- update nakadachi, query-module
+
 ## 1.2.1
 
 ### Patch Changes